+++ common/src/main/java/net/mehvahdjukaar/vista/integration/exposure/ExposureCompat.java
@@ -1,28 +1,15 @@
 package net.mehvahdjukaar.vista.integration.exposure;
 
-import io.github.mortuusars.exposure.world.item.AlbumItem;
-import io.github.mortuusars.exposure.world.item.PhotographItem;
-import io.github.mortuusars.exposure.world.item.StackedPhotographsItem;
 import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
-import net.mehvahdjukaar.moonlight.api.platform.RegHelper;
-import net.mehvahdjukaar.vista.VistaMod;
-import net.minecraft.core.component.DataComponentType;
-import net.minecraft.world.inventory.MenuType;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-
-import java.util.List;
-import java.util.function.Supplier;
 
 public class ExposureCompat {
 
     public static void init() {
-
         if (PlatHelper.getPhysicalSide().isClient()) {
             ExposureCompatClient.init();
         }
     }
-
+/*
     public static final Supplier<DataComponentType<PictureTapeContent>> PICTURE_TAPE_CONTENT =
             RegHelper.registerDataComponent(VistaMod.res("item_list"),
                     () -> DataComponentType.<PictureTapeContent>builder()
@@ -47,6 +34,6 @@ public class ExposureCompat {
         return item instanceof PictureTapeItem ||
                 item instanceof PhotographItem || item instanceof StackedPhotographsItem || item instanceof AlbumItem;
     }
-
+*/
 
 }