+++ common/src/main/java/net/mehvahdjukaar/vista/integration/exposure/PictureFrameVideoSource.java
@@ -17,6 +17,8 @@ import net.mehvahdjukaar.moonlight.api.misc.TMethod;
 import net.mehvahdjukaar.moonlight.api.util.math.Vec2i;
 import net.mehvahdjukaar.vista.client.textures.TvScreenVertexConsumers;
 import net.mehvahdjukaar.vista.client.video_source.IVideoSource;
+import net.mehvahdjukaar.vista.common.picture_tape.PictureTapeContent;
+import net.mehvahdjukaar.vista.common.picture_tape.PictureTapeItem;
 import net.mehvahdjukaar.vista.common.tv.IntAnimationState;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.resources.ResourceLocation;